backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
python-dotenv
sqlalchemy
pydantic-settings
python-magic
aiosqlite
pytest
pytest-asyncio
pytest-cov
httpx
